+++ eyg/src/eyg/ast.gleam
@@ -1,8 +1,10 @@
 import gleam/int
+import gleam/io
 import gleam/list
 import gleam/string
 import eyg/ast/provider
 import eyg/ast/expression as e
+import eyg/ast/pattern as p
 
 pub fn binary(value) {
   #(Nil, e.Binary(value))
@@ -40,7 +42,6 @@ pub fn hole() {
   provider("", e.Hole)
 }
 
-// can't use this in guards
 pub fn is_hole(generator) {
   generator == e.Hole
 }
